add secuirity action that each user just can delete own booking

+++ app/_lib/actions.js
@@ -3,6 +3,7 @@
 import { revalidatePath } from "next/cache";
 import { auth, signIn, signOut } from "./auth";
 import { supabase } from "./supabase";
+import { getBookings } from "./data-service";
 
 export async function updateGuest(formdata) {
   const session = await auth();
@@ -30,13 +31,18 @@ export async function updateGuest(formdata) {
 export async function deleteReservation(bookingId) {
   const session = await auth();
   if (!session) throw new Error("you must be logged in ");
+
+  const guestBookings = await getBookings(session.user.guestId);
+  const guestBookingIds = guestBookings.map((booking) => booking.id);
+  if (!guestBookingIds.includes(bookingId))
+    throw new Error("you are not allowed to delete this booking");
   const { error } = await supabase
     .from("bookings")
     .delete()
     .eq("id", bookingId);
 
   if (error) throw new Error("Booking could not be deleted");
-  revalidatePath("/account/reservation");
+  revalidatePath("/account/reservations");
 }
 
 export async function signInAction() {
